Changed middleware to use next-auth withAuth() wrapper
