+++ packages/app/middleware.ts
@@ -1,30 +1,32 @@
 import { getToken, JWT } from 'next-auth/jwt';
+import { withAuth } from 'next-auth/middleware';
+export { withAuth } from 'next-auth/middleware';
+
 import { NextRequest, NextResponse } from 'next/server';
 
-export async function middleware(request: NextRequest) {
-  const token = (await getToken({ req: request })) as (JWT & { user: { isGlobalAdmin: boolean } }) | null;
-  const authenticated = !!token;
+export default withAuth(
+  async function middleware(req: NextRequest) {
+    const token = (await getToken({ req })) as (JWT & { user: { isGlobalAdmin: boolean } }) | null;
 
-  const {
-    method,
-    nextUrl: { pathname },
-  } = request;
+    const pathname = req.nextUrl.pathname;
 
-  const isRead = method === 'GET';
-  const isWrite = !isRead;
-
-  let allowAccess = true;
-
-  if (pathname.startsWith('/api') && !pathname.startsWith('/api/auth')) {
-    if (pathname.startsWith('/api/profile') && isWrite) {
-      // Allow only authenticated users to write to their profile
-      allowAccess = authenticated;
-    }
-  }
-  return allowAccess ? NextResponse.next() : NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-}
+    return true ? NextResponse.next() : NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  },
+  {
+    callbacks: {
+      authorized({ token }) {
+        return token ? true : false;
+      },
+    },
+    pages: {
+      signIn: '/auth',
+    },
+  },
+);
 
 export const config = {
   // Do not run the middleware on the following paths
-  matcher: '/((?!_next/static|_next/image|manifest.json|assets|favicon*|images|logos).*)',
+  matcher: '/((?!_next/static|_next/image|manifest.json|assets|favicon*|images|logos|auth).*)',
 };
+
+// http://localhost:3000/_next/static/chunks/app/error.js
